requirements.txt
streamlit
pydrive2
pillow
pypdf2